Refazendo componentes
