+++ semana7/labefy/labefy/src/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import axios from 'axios';
-import CreatePlaylist from './components/CreatePlaylist';
-import GetAllPlaylists from './components/GetAllPlaylists';
+import CreateAndGetPlaylists from './components/CreateAndGetPlaylists';
+
 
 class App extends React.Component {
   render() {
     return (
       <div className="App">
-        <CreatePlaylist />
+        <CreateAndGetPlaylists />
       </div>
     );
   }
